Revamp repository to hold multiple JS projects
This commit is to change the structure and function of the entire repository.

I had 3 different repositories with simple JS games. So, instead of having multiple repositories with small JS projects, I want to make one collection of all my JS work. This commit is the first step to creating such a collection.